PR feedback: Catching a bug, removing unnecessary Math.floor, and modifying some of the .toDoubles

+++ core/src/test/scala/org/apache/spark/benchmark/Benchmark.scala
@@ -163,10 +163,9 @@ private[spark] class Benchmark(
     // scalastyle:on
     assert(runTimes.nonEmpty)
     val best = runTimes.min
-    val avg = runTimes.sum / runTimes.size
+    val avg = (runTimes.sum / runTimes.size).toDouble
     val stdev = if (runTimes.size > 1) {
-      math.sqrt(
-        runTimes.map(time => (time - avg) * (time - avg)).sum / (runTimes.size - 1).toDouble)
+      math.sqrt(runTimes.map(time => (time - avg) * (time - avg)).sum / (runTimes.size - 1))
     } else 0
     Result(avg / 1000000.0, num / (best / 1000.0), best / 1000000.0, stdev / 1000000.0)
   }
